refactor: create product base
